backend/requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
dj-rest-auth==1.0.6
Django==3.0.6
django-allauth==0.42.0
django-cors-headers==3.3.0
django-versatileimagefield==2.0
djangorestframework==3.11.0
idna==2.9
oauthlib==3.1.0
Pillow==7.1.2
python-magic==0.4.18
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
sqlparse==0.3.1
urllib3==1.25.9
